split maven module in 3 modules : the actual game, the DLL installer for windows, and a common library
