+++ gazeplay-commons/src/main/java/gaze/configuration/Configuration.java
