+++ gazeplay-commons/src/main/java/utils/games/Utils.java
@@ -1,25 +1,13 @@
 package utils.games;
 
-import javafx.event.Event;
-import javafx.event.EventHandler;
-import javafx.scene.Group;
-import javafx.scene.Node;
-import javafx.scene.Scene;
-import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
 import javafx.scene.control.MenuItem;
 import javafx.scene.image.Image;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.media.Media;
 import javafx.scene.media.MediaPlayer;
-import javafx.scene.paint.Color;
-import javafx.scene.paint.ImagePattern;
-import javafx.scene.shape.Rectangle;
 import lombok.extern.slf4j.Slf4j;
-import org.tc33.jheatchart.HeatChart;
-import utils.games.stats.Stats;
-import utils.games.stats.StatsDisplay;
+import org.apache.commons.io.IOUtils;
 
 import javax.activation.MimetypesFileTypeMap;
 import java.io.*;
@@ -36,13 +24,13 @@ import java.util.Date;
 public class Utils {
 
     public static final String FILESEPARATOR = System.getProperties().getProperty("file.separator");
-    private static final String heatMapFileName = "java-heat-chart.png";
+
 
     private static final String tempFolder = "temp";
 
     private static String filesFolder = getGazePlayFolder() + "files" +FILESEPARATOR;
 
-    public static Home home;
+
 
     public static Image[] getImages(String folder) {
 
@@ -99,107 +87,8 @@ public class Utils {
         return mimetype.startsWith("image/");
     }
 
-    public static void clear(Scene scene, Group root, ChoiceBox<String> cbxGames) {
-
-        scene.setFill(Color.BLACK);
-
-        for (Node N : root.getChildren()) {
-
-            N.setTranslateX(-10000);
-        }
-
-        root.getChildren().remove(0, root.getChildren().size());
-
-        root.getChildren().add(Bravo.getBravo());
 
 
-    }
-
-    public static void home(Scene scene, Group root, ChoiceBox<String> cbxGames, Stats stats) {
-
-        double width = scene.getWidth() / 10;
-        double height = width;
-        double X = scene.getWidth() * 0.9;
-        double Y = scene.getHeight() - height * 1.1;
-
-        home = new Home(X, Y, width, height);
-
-        EventHandler<Event> homeEvent = new EventHandler<javafx.event.Event>() {
-            @Override
-            public void handle(javafx.event.Event e) {
-
-                if (e.getEventType() == MouseEvent.MOUSE_CLICKED) {
-
-                    log.info("stats = " + stats);
-
-                    if (stats == null) {
-
-                        goHome(scene, root, cbxGames);
-                    }
-                    else{
-
-                        StatsDisplay.displayStats(stats, scene, root, cbxGames);
-                    }
-                }
-            }
-        };
-
-        home.addEventHandler(MouseEvent.MOUSE_CLICKED, homeEvent);
-
-        root.getChildren().add(home);
-    }
-
-    private static void goHome(Scene scene, Group root, ChoiceBox<String> cbxGames) {
-
-        clear(scene, root, cbxGames);
-
-        if (cbxGames != null) {
-            cbxGames.getSelectionModel().clearSelection();
-            root.getChildren().add(cbxGames);
-
-            cbxGames.setTranslateX(scene.getWidth() * 0.9 / 2);
-            cbxGames.setTranslateY(scene.getHeight() * 0.9 / 2);
-
-            addButtons(scene, root, cbxGames);
-        }
-    }
-
-    public static void addButtons(Scene scene, Group root, ChoiceBox<String> cbxGames) {
-
-        double width = scene.getWidth() / 10;
-        double heigth = width;
-        double XExit = scene.getWidth() * 0.9;
-        double XLicence = scene.getWidth() * 0.0;
-        double Y = scene.getHeight() - heigth * 1.1;
-
-        //  License license = new License(XLicence, Y, width, heigth, scene, root, cbxGames);
-
-        // root.getChildren().add(license);
-
-        Rectangle exit = new Rectangle(XExit, Y, width, heigth);
-
-        exit.setFill(new ImagePattern(new Image("data/common/images/power-off.png"), 0, 0, 1, 1, true));
-
-        EventHandler<javafx.event.Event> homeEvent = new EventHandler<javafx.event.Event>() {
-            @Override
-            public void handle(javafx.event.Event e) {
-
-                if (e.getEventType() == MouseEvent.MOUSE_CLICKED) {
-
-                    System.exit(0);
-
-                }
-            }
-        };
-
-        exit.addEventHandler(MouseEvent.MOUSE_CLICKED, homeEvent);
-
-        root.getChildren().add(exit);
-
-        root.getChildren().add(logo(scene));
-
-        root.getChildren().add(Utils.BuildLicence());
-    }
 
     public static MenuBar BuildLicence() {
 
@@ -235,19 +124,6 @@ public class Utils {
         return menuBar;
     }
 
-    public static Node logo(Scene scene) {
-
-        double width = scene.getWidth()*0.5;
-        double height = scene.getHeight()*0.2;
-
-        double posY = scene.getHeight()*0.1;
-        double posX = (scene.getWidth() - width)/2;
-
-        Rectangle logo = new Rectangle(posX,posY, width,height);
-        logo.setFill(new ImagePattern(new Image("data/common/images/gazeplay.jpg"),0,0,1,1, true));
-
-        return logo;
-    }
 
     public static Image[] images(String folder){
 
@@ -431,36 +307,8 @@ public class Utils {
         }
     }
 
-    public static void buildHeatMap(double[][] data){
 
-        // Step 1: Create our heat map chart using our data.
-        HeatChart map = new HeatChart(data);
 
-        map.setHighValueColour(java.awt.Color.RED);
-        map.setLowValueColour(java.awt.Color.lightGray);
-
-        map.setShowXAxisValues(false);
-        map.setShowYAxisValues(false);
-        map.setChartMargin(0);
-
-        //creation of temp folder if it doesn't already exist.
-        (new File(Utils.getTempFolder())).mkdir();
-
-        String heatMapPath = Utils.getTempFolder() + heatMapFileName;
-
-        File saveFile = new File(heatMapPath);
-
-        try {
-            map.saveToFile(saveFile);
-        } catch (IOException e) {
-            log.error("Exception", e);
-        }
-    }
-
-    public static String getHeatMapPath(){
-
-        return Utils.getTempFolder() + heatMapFileName;
-    }
 
     public static PrintWriter getInfoStatsFile(String folder) {
 
@@ -476,21 +324,23 @@ public class Utils {
         return out;
     }
 
-    public static boolean copyFromJar(String FilePath, String destinationPath){
-
-        try{
-            InputStream sourceFile = ClassLoader.getSystemResourceAsStream(FilePath);
-            OutputStream destinationFile = new FileOutputStream(destinationPath);
-            // Lecture par segment de 0.5Mo
-            byte buffer[] = new byte[512 * 1024];
-            int nbLecture;
-            while ((nbLecture = sourceFile.read(buffer)) != -1){
-                destinationFile.write(buffer, 0, nbLecture);
+    public static boolean copyFromJar(String filePath, String destinationPath){
+		InputStream sourceFile = null;
+		OutputStream destinationFile = null;
+		try{
+            sourceFile = ClassLoader.getSystemResourceAsStream(filePath);
+            if (sourceFile == null) {
+                throw new IOException("Resource not found " + filePath);
             }
+            destinationFile = new FileOutputStream(destinationPath);
+            org.apache.commons.io.IOUtils.copy(sourceFile, destinationFile);
         } catch (IOException e){
             log.error("Exception", e);
             return false; // Erreur
-        }
+        } finally {
+			IOUtils.closeQuietly(destinationFile);
+			IOUtils.closeQuietly(sourceFile);
+		}
         return true; // Résultat OK
     }
 }
