+++ pom.xml
@@ -14,7 +14,9 @@
 	</properties>
 
 	<modules>
+		<module>gazeplay-commons</module>
 		<module>gazeplay</module>
+		<module>gazeplay-tobii-setup</module>
 	</modules>
 
 
@@ -90,6 +92,12 @@
 				<artifactId>lombok</artifactId>
 				<version>1.16.18</version>
 				<scope>provided</scope>
+			</dependency>
+
+			<dependency>
+				<groupId>commons-io</groupId>
+				<artifactId>commons-io</artifactId>
+				<version>2.5</version>
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
